Move slow newapp tests to integration
